Adds support for query persistence in local storage
