+++ package.json
@@ -78,8 +78,10 @@
     "@radix-ui/react-tooltip": "^1.2.0",
     "@sentry/browser": "^9.0.0",
     "@tailwindcss/vite": "^4.0.17",
+    "@tanstack/query-async-storage-persister": "^5.83.0",
     "@tanstack/react-query": "^5.64.1",
     "@tanstack/react-query-devtools": "^5.64.2",
+    "@tanstack/react-query-persist-client": "^5.83.0",
     "@vitejs/plugin-react": "^4.3.4",
     "bowser": "^2.11.0",
     "browser-image-compression": "^2.0.2",
